# Optimize parging option args