@ohler55 2021-08-07
Can we count on the symbols always being the same? In other words, could they change or all they always interned?