@Watson1978 2021-08-07
Unlike objects such as strings, the same symbol will be assigned to a unique address.
Therefore, the id of the Symbol object will also be unique.

```
irb(main):001:0> "foo".object_id
=> 260
irb(main):002:0> "foo".object_id
=> 280
irb(main):003:0> :foo.object_id
=> 2192668
irb(main):004:0> :foo.object_id
=> 2192668
```

However, it may be safer to compare the values converted by SYM2ID().