@ohler55 2021-08-07
If IDs are assured to be the same when the content is the same for the life of the program then that might be better.