@Watson1978 2021-08-07
Since we use rb_gc_register_address() to register the symbol, the symbol will not be freed.

Here is one of using `rb_gc_register_address()`.
https://github.com/ohler55/oj/blob/62aea593dcd7c2102600eeb8dd244f1b39e4088d/ext/oj/oj.c#L1881

OK, I will update PR to use IDs.