@ohler55 2021-08-07
As long as the symbols or IDs don't change I'm okay with either one.