@Watson1978 2021-08-07
> As long as the symbols or IDs don't change I'm okay with either one.

Thanks. I'm going to quit updating to use IDs.
We have already compared Symbol directly in several places.

https://github.com/ohler55/oj/blob/62aea593dcd7c2102600eeb8dd244f1b39e4088d/ext/oj/oj.c#L1063-L1076