@Watson1978 2021-08-07
Ops, I found the mistake in `Test code` in description.

- ❌`Oj.dump(json, mode: :compat)`
- ✅`Oj.dump(data, mode: :compat)`

I should pass `data` instead of `json` string...