@Watson1978 2021-08-07
OK, I retake the result.


−               | before   | after    | result
--               | --       | --       | --
Oj.load          | 387.405k | 420.518k | 1.09x
Oj.dump          | 601.762k | 705.901k | 1.17x

### Before
```
Warming up --------------------------------------
             Oj.load    38.675k i/100ms
             Oj.dump    60.591k i/100ms
Calculating -------------------------------------
             Oj.load    387.405k (± 0.9%) i/s -      1.972M in   5.091756s
             Oj.dump    601.762k (± 0.7%) i/s -      3.030M in   5.034717s
```

### After
```
Warming up --------------------------------------
             Oj.load    42.372k i/100ms
             Oj.dump    70.934k i/100ms
Calculating -------------------------------------
             Oj.load    420.518k (± 0.3%) i/s -      2.119M in   5.038126s
             Oj.dump    705.901k (± 0.3%) i/s -      3.547M in   5.024398s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

Benchmark.ips do |x|
  x.report('Oj.load (symbol_keys)') { Oj.load(json, symbol_keys: true) }

  data = Oj.load(json, symbol_keys: true)
  x.report('Oj.dump (compat)') { Oj.dump(data, mode: :compat) }
end
```